removed Singleton from Interactor model
Another way of parallelization was made with id of the object and
ctypes. It is a code more dangerous but if the object remains in the
memory during the function execution nothing can go wrong, but this
assumption need to prevail.

The benefit of this change is that the model is free from
Singleton. It is more object driven.